Revert "update 27.12.2014"
This reverts commit 14ca55e1976809c52ac0137bc4633468f0f1e6a1.

Conflicts:
	src/include/gld_drawing.cpp
	src/include/gld_drawing.h
	src/include/gld_visual_view.cpp
	src/include/gld_visual_view.h
	src/include/glv_visual_model.cpp
	src/include/glv_visual_model.h

+++ src/include/gld_interface.cpp
@@ -58,7 +58,7 @@ void Button::draw(){
 			glColor4fv(color[TEXT_CLICKED]);
 		else if (border.contains(input.mouse.x, input.mouse.y) && color[TEXT_HOVERED] != NULL)
 			glColor4fv(color[TEXT_HOVERED]);
-		font->textOut(border.left + BUTTON_PADDING, border.top + ( border.height - (int)(font->getHeight()*0.7) ) / 2, textOnButton);
+		font->textOut(border.left + BUTTON_PADDING, border.top + ( border.height - font->getHeight()*0.7 ) / 2, textOnButton);
 	}
 	if (isSeparator()) {
 		glColor4fv(color[BORDER_NORMAL]);
